♻️ Refactor create and edit comment with custom hooks
